requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
psycopg2-binary>=2.9
drf-spectacular>=0.27.0
python-dotenv>=1.0
dj-database-url>=3.0.1